+++ src/cmds/exec.js
@@ -30,6 +30,11 @@ export default {
           describe: '',
           default: userConfig.exec.bail
         },
+        prefix: {
+          type: 'boolean',
+          describe: 'Prefix output with the package name',
+          default: userConfig.exec.prefix
+        },
         concurrency: {
           type: 'number',
           describe: 'How many commands to run at the same time',