+++ src/types.ts
@@ -338,6 +338,11 @@ interface ReleaseRcOptions {
   tag: string
 
   /**
+   * Prefix output with the package name
+   */
+  prefix?: boolean
+
+  /**
    * Release modules in parallel up to this limit
    */
   concurrency?: number
@@ -362,6 +367,11 @@ interface ExecOptions {
   bail?: boolean
 
   /**
+   * Prefix output with the package name
+   */
+  prefix?: boolean
+
+  /**
    * Run commands in parallel up to this limit
    */
   concurrency?: number
@@ -372,6 +382,11 @@ interface RunOptions {
    * If false, the script will continue to be run in other packages
    */
   bail?: boolean
+
+  /**
+   * Prefix output with the package name
+   */
+  prefix?: boolean
 
   /**
    * Run scripts in parallel up to this limit